uniforms are working

+++ src/swapchain.h
@@ -33,7 +33,7 @@ public:
 	inline VkFormat getFormat() { return format; }
 	inline VkExtent2D getExtent() { return extent; }
 	inline uint32_t getImageCount() { return static_cast<uint32_t>(image_views.size()); }
-	inline VkImageView getImage(uint32_t i) { return image_views[i]; }
+	inline VkImageView getImage(size_t i) { return image_views[i]; }
 	inline VkSwapchainKHR getSwapchain() { return swapchain; }
 
 	static SwapchainSupportInfo getSupportInfo(VkPhysicalDevice device, VkSurfaceKHR surface);
